* [FIX] Handle ISE generated by locale issues on php8, on Tiki 26.
* https://avan.tech/item93949
* https://dev.tiki.org/item8448
On OSX, MAMP (PHP 8.1) setlocale(LC_ALL, '') was leading to an Internal Server Error (ISE) in tiki-setup.php.
Now Tiki checks first the system's default locale to make sure that it is supported by php or is valid and that
it is installed and then set it the current locale.