[NEW] Add Composer Version Validation to Ensure Compatibility with Composer 2 --- * [NEW] Add Composer Version Validation to Ensure Compatibility with Composer 2
Introduces a validation mechanism to ensure that Tiki Manager use Composer 2 or higher, preventing issues caused by running with Composer 1 or outdated versions of Composer.
Key Changes:

Updated composer.json to include a pre-install-cmd and pre-update-cmd scripts that check the version of Composer before running installation commands.
Implemented a ValidateVersion class that compares the running Composer version with the required minimum (Composer 2.0.0) and halts execution if the version is not supported.
Added an exception message that will alert users if they are attempting to use Composer 1 or an outdated version of Composer 2, ensuring they upgrade to the appropriate version.

See merge request tikiwiki/tiki!5995